Move remaining `load` modules
