+++ crates/cubecl-attention/src/components/global/dummy/read.rs
