+++ crates/cubecl-convolution/src/components/global/mod.rs
@@ -1,9 +1,9 @@
 pub mod args;
 pub mod entry_point;
 pub mod layout;
-pub mod load;
 pub mod memory;
 pub mod multi_stage;
+pub mod read;
 pub mod single_stage;
 
 mod base;
