+++ crates/cubecl-convolution/src/components/global/read/mod.rs
