+++ crates/cubecl-convolution/src/components/global/read/reader/mod.rs
